+++ kion/provider_test.go
@@ -1,0 +1,37 @@
+package kion
+
+import (
+	"os"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+var testAccProviders map[string]*schema.Provider
+var testAccProvider *schema.Provider
+
+func init() {
+	testAccProvider = Provider()
+	testAccProviders = map[string]*schema.Provider{
+		"kion": testAccProvider,
+	}
+}
+
+func TestProvider(t *testing.T) {
+	if err := Provider().InternalValidate(); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+}
+
+func TestProvider_impl(t *testing.T) {
+	var _ *schema.Provider = Provider()
+}
+
+func testAccPreCheck(t *testing.T) {
+	if err := os.Getenv("KION_URL"); err == "" {
+		t.Fatal("KION_URL must be set for acceptance tests")
+	}
+	if err := os.Getenv("KION_APIKEY"); err == "" {
+		t.Fatal("KION_APIKEY must be set for acceptance tests")
+	}
+}